// src/api/ak-ambulance-counseling-wl/models/LoginForm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Waiting List Api
 * Ambulance Counseling Project API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xkoricansky@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginForm
 */
export interface LoginForm {
    /**
     * Email address of the user
     * @type {string}
     * @memberof LoginForm
     */
    email: string;
    /**
     * Password for the user account
     * @type {string}
     * @memberof LoginForm
     */
    password: string;
}

/**
 * Check if a given object implements the LoginForm interface.
 */
export function instanceOfLoginForm(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "password" in value;

    return isInstance;
}

export function LoginFormFromJSON(json: any): LoginForm {
    return LoginFormFromJSONTyped(json, false);
}

export function LoginFormFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginForm {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'email': json['email'],
        'password': json['password'],
    };
}

export function LoginFormToJSON(value?: LoginForm | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'email': value.email,
        'password': value.password,
    };
}