// src/api/ak-ambulance-counseling-wl/models/RegistrationForm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Waiting List Api
 * Ambulance Counseling Project API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xkoricansky@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RegistrationForm
 */
export interface RegistrationForm {
    /**
     * Name of the user
     * @type {string}
     * @memberof RegistrationForm
     */
    name: string;
    /**
     * Email address of the user
     * @type {string}
     * @memberof RegistrationForm
     */
    email: string;
    /**
     * Password for the user account
     * @type {string}
     * @memberof RegistrationForm
     */
    password: string;
}

/**
 * Check if a given object implements the RegistrationForm interface.
 */
export function instanceOfRegistrationForm(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "password" in value;

    return isInstance;
}

export function RegistrationFormFromJSON(json: any): RegistrationForm {
    return RegistrationFormFromJSONTyped(json, false);
}

export function RegistrationFormFromJSONTyped(json: any, ignoreDiscriminator: boolean): RegistrationForm {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'email': json['email'],
        'password': json['password'],
    };
}

export function RegistrationFormToJSON(value?: RegistrationForm | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'email': value.email,
        'password': value.password,
    };
}